unsigned long => unsigned int

+++ src/fancy/IMagickImage.cpp
@@ -229,11 +229,11 @@ Magick::Color IMagickImage::process_element(unsigned int x0, unsigned int y0,
                                             unsigned int x1, unsigned int y1)
 {
   
-  unsigned long r = 0, 
+  unsigned int r = 0, 
                 g = 0, 
                 b = 0;
                 
-  unsigned long total = 0;
+  unsigned int total = 0;
   
   Magick::Color color;
